satellite/cosole: add tests for wasm access code
